p1

+++ include/ast.hh
@@ -50,9 +50,9 @@ struct NodeBinOp : public Node {
     Node for integer literals
 */
 struct NodeInt : public Node {
-    int value;
+    long value;
 
-    NodeInt(int val);
+    NodeInt(long val);
     std::string to_string();
     llvm::Value *llvm_codegen(LLVMCompiler *compiler);
 };
